requirements.txt
pywin32>=306
PyQt6>=6.6
img2pdf>=0.5.1
Pillow>=10.2
